+++ thursday/week2/laziness/Lazy1427.java
@@ -1,6 +1,6 @@
 import java.io.*;
 
-public class Main {
+public class Lazy1427 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         char[] arr = br.readLine().toCharArray();
